+++ src/shapes/Square.java
@@ -2,16 +2,16 @@ package shapes;
 
 public class Square extends Quadrilateral {
 
-    public Square(int side) {
+    public Square(double side) {
         super(side, side);
     }
 
-    public void setLength (int length){
+    public void setLength (double length){
         this.length = length;
         this.width = length;
     }
 
-    public void setWidth (int width){
+    public void setWidth (double width){
         this.width = width;
         this.length = width;
     }
